feat(anytls): add dedicated AnyTLS outbound with native TCP/UDP and robust datagram framing (#1317)
* Initial plan

* feat: add anytls outbound support via trojan-compatible handler

Agent-Logs-Url: https://github.com/Watfaq/clash-rs/sessions/93c58896-abeb-48aa-839c-04f36b2e9398


* chore: document and warn for unsupported anytls optional fields

Agent-Logs-Url: https://github.com/Watfaq/clash-rs/sessions/93c58896-abeb-48aa-839c-04f36b2e9398


* refactor: use dedicated anytls handler namespace in load paths

Agent-Logs-Url: https://github.com/Watfaq/clash-rs/sessions/8b36e420-d314-4d89-b724-8e584bfbd471


* feat: implement dedicated anytls outbound handler module

Agent-Logs-Url: https://github.com/Watfaq/clash-rs/sessions/cc66761b-35a2-4d7c-8a9e-3b97a017e7f4


* fix: simplify anytls datagram read_exact handling

Agent-Logs-Url: https://github.com/Watfaq/clash-rs/sessions/cc66761b-35a2-4d7c-8a9e-3b97a017e7f4


* feat: rework anytls outbound to use native anytls framing

Agent-Logs-Url: https://github.com/Watfaq/clash-rs/sessions/86e2d767-afaf-4b41-9a94-494c8012061a


* test: add anytls docker test and restore trojan behavior

Agent-Logs-Url: https://github.com/Watfaq/clash-rs/sessions/046efdb8-7f5c-42a9-a2f5-ed2c7902db7a


* fix: address clippy findings and stabilize docker-test expectations

Agent-Logs-Url: https://github.com/Watfaq/clash-rs/sessions/a4b776ad-1ea3-4d9b-888f-7c3d985db84a


* feat: add anytls udp-over-tcp v2 outbound support

Agent-Logs-Url: https://github.com/Watfaq/clash-rs/sessions/0b2d2bb4-1258-4e76-b0ad-0249372d9505


* fix: make anytls udp datagram reads packet-safe

Agent-Logs-Url: https://github.com/Watfaq/clash-rs/sessions/0b2d2bb4-1258-4e76-b0ad-0249372d9505


* revert: restore trojan docker test suite coverage

Agent-Logs-Url: https://github.com/Watfaq/clash-rs/sessions/08d820aa-cd02-442a-b06f-5167d12a99b1


* fix(anytls): address PR review comments - fix Poll::Pending deadlock in datagram, add CancellationToken for relay shutdown


* test(anytls): add docker test coverage with xray anytls server config

- Update anytls.json to use 'accounts' key and explicit tcp network setting
  to match the xray anytls inbound protocol schema
- Remove #[ignore] from test_anytls_tcp; use Suite::tcp_tests() for full
  TCP coverage (PingPongTcp + LatencyTcp)
- Add test_anytls_udp with udp: true and Suite::all() to cover UDP-over-TCP
  via the UoT v2 mechanism


* test(anytls): merge tcp and udp docker tests into single test_anytls with Suite::all()


* refactor(anytls): code quality improvements and simplification

- Move CMD_* and STREAM_ID constants to module level so they are
  visible across the module and not buried inside a method body
- Add biased; to both relay select! loops to ensure the CancellationToken
  branch is always checked first, guaranteeing prompt shutdown
- Remove 64 KB read_buf scratch buffer from OutboundDatagramAnytls;
  read UDP payload directly into packet_buf spare capacity using
  ReadBuf::uninit + BytesMut::advance_mut, eliminating one copy per
  received datagram and one 65535-byte per-connection allocation
- Simplify ALPN default collection: DEFAULT_ALPN.map(str::to_owned).to_vec()
- Simplify SNI clone: .clone().unwrap_or_else(|| …clone()) avoids
  going through &String -> String and is the idiomatic Option idiom


* fix(anytls): allow private IP routing in xray test config and expand unit test coverage

- Add explicit geoip:private routing rule to anytls.json so xray does not
  apply its default private-IP blocking policy for the anytls inbound
- Add 10 new unit tests covering Handler proto/name/udp flag/as_map,
  datagram write-length-prefix, roundtrip, and oversized-packet rejection


* fix(anytls): switch to sing-box docker image and expand test coverage

- Replace xray with ghcr.io/sagernet/sing-box:v1.13.8 for docker test
  (xray-core does not support anytls inbound; sing-box supports it since v1.12.0)
- Convert anytls.json to sing-box config format (listen_port, users, tls)
- Mount config to /etc/sing-box/config.json
- Add IMAGE_SINGBOX constant to docker test consts
- Merge unit_tests and docker tests into single mod tests block
- Add tests for open_anytls_stream: handshake bytes, relay data path
- Add tests for inner_proxy_stream (no TLS/transport path)
- Add write_frame/read_frame roundtrip and empty-payload tests
- Move PlainProxyAPIResponse impl before mod tests (fixes items_after_test_module lint)


* fix(tests): switch anytls to sing-box, pin xray to 26.3.27, add unit tests

- Replace xray docker image with ghcr.io/sagernet/sing-box:v1.13.8 for anytls
  (xray-core has never supported anytls inbound)
- Add .cmd(["run", "-c", "/etc/sing-box/config.json"]) so sing-box starts
- Rewrite anytls.json in sing-box format with anytls inbound + tls
- Pin IMAGE_XRAY to teddysun/xray:26.3.27 — v26.4.17 blocks private IPs
  for trojan inbound, breaking test_trojan_grpc on all PRs
- Add IMAGE_SINGBOX constant to docker test consts
- Merge unit_tests and docker tests modules into single mod tests
- Add 5 unit tests covering open_anytls_stream, inner_proxy_stream,
  write_frame/read_frame roundtrip


* fix(anytls): bump to protocol v2, add padding-md5, fix padding scheme

- Bump cmdSettings version v=1 → v=2 (current protocol version)
- Add padding-md5 to cmdSettings (MD5 of our "stop=0" scheme)
- Add CMD_SERVER_SETTINGS (10) and CMD_UPDATE_PADDING_SCHEME (6) constants
  and handle them in the reader task (read and discard)
- Add "padding_scheme": ["stop=0"] to sing-box server test config so the
  server does not enforce padding length on auth — required because the
  default scheme (0=30-30) rejects clients that send 0 bytes of padding
- Update unit test to assert v=2 and presence of correct padding-md5


* fix(anytls): batch entire handshake into single write_all for one TLS record

tokio-rustls creates a separate TLS application-data record per
poll_write call. sing-box reads the first record as a complete auth
packet, so splitting the 32-byte hash + u16(padding_len) across two
separate write calls caused sing-box to get EOF while reading the
padding-length field.

Fix: accumulate hash + padding_len + cmdSettings + cmdSYN + cmdPSH
into a single BytesMut buffer and call write_all once, guaranteeing
a single TLS record.


---------

+++ clash-lib/src/proxy/group/smart/stats.rs
@@ -96,7 +96,7 @@ impl SiteStats {
 
         for delay in self.delay_history.iter() {
             // Use the calculated age for time weighting
-            let time_weight = (-0.1 * age_secs as f64).exp(); // Specify f64
+            let time_weight = (-0.1 * age_secs).exp();
             let delay_weight = (-0.001 * *delay).exp(); // Dereference delay, remove redundant cast. Higher delays get less weight
             let weight = time_weight * delay_weight;
